- Mosquitto Broker functional
