+++ WebApp/yii-application/composer.json
@@ -24,7 +24,8 @@
         "mpdf/mpdf": "^8.2",
         "dompdf/dompdf": "^2.0",
         "2amigos/qrcode-library": "3.1.0",
-        "symfony/var-dumper": "^5.4"
+        "symfony/var-dumper": "^5.4",
+        "bluerhinos/phpmqtt": "@dev"
     },
     "require-dev": {
         "yiisoft/yii2-debug": "~2.1.0",
